pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minifw-ai"
version = "2.2.0"
description = "MiniFW-AI Firewall with AI detection"
authors = [{name = "MiniFW-AI Team"}]
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "python-dotenv",
    "jinja2",
    "python-multipart",
    "pydantic",
    "pydantic-settings",
    "bcrypt",
    "pyjwt",
    "apscheduler",
    "prometheus-client",
    "scikit-learn",
    "pandas",
    "numpy"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "minifw_ai*"]   # hanya dua package ini
exclude = ["logs*", "config*", "models*", "testing*", "systemd*", "prometheus*", "scheduler*", "yara_rules*"]
